Fix Docker mount bug
For the tests, we mount a lot of filesystem images, which entails attaching the file to a loop device (/dev/loopX) and mounting the loop device. If every existing loop device is already in use, a new loop device is created.

Our Docker test container has access to the devices from the host system (it is allowed to via the --privileged flag). However, it looks like it can only see the /dev directory as it was when the container was started. Therefore, the test container worked fine until it ran out of loop devices, at which point it could not create any new loop devices and tests started failing en masse. We fix this most annoying of Heisenbugs by mounting the host's /dev directory as a volume.